Support URL's with containing whitespace
Improve support for checking URL's containing whitepace by unescaping
the URL and then removing tabs and CR/LF characters.  This allows
URL's such as the one below to be checked correctly:

<a href="/path/to/page/i-am-a-really-long-
title-that-got-wrapped">Really Long Title</a>

Browsers such as Chrome and Firefox remove the embedded whitespace
resulting in the URL
'/path/to/page/i-am-a-really-long-title-that-got-wrapped'.  This
commit tries to duplicate this behavior.