Add support for ini file.
Motivation:

In CI we have multiple invocations of muffet with mostly the same arguments.
This is cumbersome to maintain all the other args.
Thus we'd like to have support for a config file to use as default

Implementation:

Always use `muffet.ini`. I thought it was likely unecessary to support
anything different until the need comes for it.

Signed-off-by: Charly Molter <charly.molter@konghq.com>