Added --ignore-network-errors option to optionally ignore network errors
The idea of the option is similar to --accepted-status-codes=200..300,403,429,500..599
It prevents false link check errors when some remote host is temporarily down
or times out a request (often happens when there are lots of external links),
but still allows checking for 404 when the remote host is functioning.

The default behavior is not to ignore these errors, as it was before. It's possible
to either ignore all network errors, or to only ignore errors on external hosts.